% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ph_init}
\alias{ph_init}
\title{Page–Hinkley Change Detector: Initialize State}
\usage{
ph_init(delta = 0.005, lambda = 50, alpha = 0.999)
}
\arguments{
\item{delta}{Insensitivity parameter for mean deviation.}

\item{lambda}{Detection threshold.}

\item{alpha}{Exponential moving average factor for the mean.}
}
\value{
List state to be passed to \code{\link{ph_update}}.
}
\description{
\code{ph_init} initializes the internal state for the Page–Hinkley change
detection test with exponential moving average.
}
\examples{
\dontrun{
s <- ph_init()
}
}
\seealso{
Other viz-monitoring: 
\code{\link{ph_update}()},
\code{\link{plot_data_comparison}()}
}
\concept{viz-monitoring}
